initial tensor hook refactor
the grad passed into tensor hooks does not compare with the recomputed grad, but the model still converges. The full_backward_hook also seems to be incompatible with tensor hooks, as it is called after the tensor hooks are called. This was not a problem with the now deprecated hooks. I think this approach is a bust.